// src/lib/generated/google-fonts.ts
export interface GoogleFontMeta {
	category: 'sans' | 'serif' | 'mono';
	axes: { tag: string; min: number; max: number }[];
	weights: number[];
	subsets: string[];
	variable: boolean;
}